split files and move to folders
