+++ js/classes/player-answer-class.js
@@ -1,0 +1,8 @@
+class PlayerAnswer {
+  constructor(boolean, time = 30000) {
+    this.isRightAnswer = boolean;
+    this.time = time;
+  }
+}
+
+export default PlayerAnswer;
