+++ js/templates/attempts-ended-screen.js
@@ -1,7 +1,7 @@
-import getHtmlFromTemplate from './get-html-from-template';
+import getHtmlFromTemplate from '../get-html-from-template';
 import {getWelcomeScreenMarkup} from './welcome-screen';
-import renderScreen from './render-screen';
-import {initialState} from "./data/data";
+import renderScreen from '../render-screen';
+import {initialState} from "../data/initial-data";
 
 const attemptsEndedScreenMarkup = getHtmlFromTemplate(`<section class="main main--result">
   <section class="logo" title="Угадай мелодию"><h1>Угадай мелодию</h1></section>
