+++ js/templates/level-genre-screen.js
@@ -1,9 +1,10 @@
-import getHtmlFromTemplate from './get-html-from-template';
+import getHtmlFromTemplate from '../get-html-from-template';
 import {currentState} from './welcome-screen';
-import {games, PlayerAnswer} from './data/data';
-import getGameHeaderTemplate from './view/game-header';
-import {checkGenre} from './utils';
-import switchScreen from './switch-screens';
+import {games} from '../data/initial-data';
+import getGameHeaderTemplate from '../views/game-header';
+import {checkGenre} from '../utils';
+import switchScreen from '../switch-screens';
+import PlayerAnswer from '../classes/player-answer-class';
 
 /**
  * On answer button click handler
