+++ js/templates/time-out-screen.js
@@ -1,4 +1,4 @@
-import getHtmlFromTemplate from './get-html-from-template';
+import getHtmlFromTemplate from '../get-html-from-template';
 
 const timeOutScreenMarkup = getHtmlFromTemplate(`<section class="main main--result">
   <section class="logo" title="Угадай мелодию"><h1>Угадай мелодию</h1></section>
