+++ js/views/game-header.js
